backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
chromadb==0.4.18
pypdf==3.17.4
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
google-cloud-translate==3.13.0
sentence-transformers==2.2.2
openai==1.6.1
pydantic==2.5.2
ollama==0.1.7
langchain-ollama==0.0.1